Added ConfigurationManager and removed old method from Http managers
