+++ SteelSeriesAPI/Sonar/Http/SonarHttpProvider.cs
@@ -22,88 +22,6 @@ public class SonarHttpProvider : ISonarDataProvider
         return (Mode)ModeExtensions.FromDictKey(mode, ModeMapChoice.StreamDict);
     }
 
-    #region AudioConfigs
-    
-    public IEnumerable<SonarAudioConfiguration> GetAllAudioConfigurations()
-    {
-        JsonDocument configs = new HttpProvider("configs").Provide();
-
-        foreach (var element in configs.RootElement.EnumerateArray())
-        {
-            string vDevice = element.GetProperty("virtualAudioDevice").GetString();
-            string id = element.GetProperty("id").GetString();
-            string name = element.GetProperty("name").GetString();
-            
-            yield return new SonarAudioConfiguration(id, name, (Channel)ChannelExtensions.FromDictKey(vDevice));
-        }
-    }
-
-    public SonarAudioConfiguration GetAudioConfiguration(string configId)
-    {
-        IEnumerable<SonarAudioConfiguration> configs = GetAllAudioConfigurations();
-        SonarAudioConfiguration sonarConfig = null;
-        
-        foreach (var config in configs)
-        {
-            if (config.Id == configId)
-            {
-                sonarConfig = config;
-                break;
-            }
-        }
-
-        return sonarConfig;
-    }
-    
-    public IEnumerable<SonarAudioConfiguration> GetAudioConfigurations(Channel channel)
-    {
-        if (channel == Channel.MASTER)
-        {
-            throw new Exception("Can't get audio configurations for master");
-        }
-        
-        IEnumerable<SonarAudioConfiguration> configs = GetAllAudioConfigurations();
-        List<SonarAudioConfiguration> deviceConfigs = new List<SonarAudioConfiguration>();
-        
-        foreach (var config in configs)
-        {
-            if (config.AssociatedChannel == channel)
-            {
-                deviceConfigs.Add(config);
-            }
-        }
-
-        return deviceConfigs.OrderBy(s => s.Name);
-    }
-    
-    public SonarAudioConfiguration GetSelectedAudioConfiguration(Channel channel)
-    {
-        if (channel == Channel.MASTER)
-        {
-            throw new Exception("Can't get audio configuration for master");
-        }
-
-        JsonDocument selectedConfigs = new HttpProvider("configs/selected").Provide();
-        JsonElement sConfig = default;
-
-        foreach (var config in selectedConfigs.RootElement.EnumerateArray())
-        {
-            if (config.GetProperty("virtualAudioDevice").GetString() == channel.ToDictKey())
-            {
-                sConfig = config;
-                break;
-            }
-        }
-
-        string id = sConfig.GetProperty("id").GetString();
-        string name = sConfig.GetProperty("name").GetString();
-        string vDevice = sConfig.GetProperty("virtualAudioDevice").GetString();
-
-        return new SonarAudioConfiguration(id, name, (Channel)ChannelExtensions.FromDictKey(vDevice));
-    }
-    
-    #endregion
-
     #region ChatMix
 
     public double GetChatMixBalance()
